Run bookshop model in worker too
